+++ src/hooks/useBookmark.js
@@ -1,0 +1,37 @@
+import axios from "axios";
+import { useState } from "react";
+
+import { useAuthContext } from "../context/AuthContext";
+
+export default function useBookmark() {
+  const [isDeleted, setIsDeleted] = useState(false);
+  const { currentUser } = useAuthContext();
+
+  const setIdToken = async () => {
+    const token = await currentUser?.getIdToken();
+    return token;
+  };
+
+  const deleteBookmark = async (id) => {
+    const token = await setIdToken();
+    if (!token) {
+      throw new Error('No token found');
+    }
+    const config = {
+      headers: { authorization: `Bearer ${token}` },
+    };
+    const res = await axios.delete(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/bookmarks/${id}`, config);
+    if (res.status === 204) {
+      setIsDeleted(true);
+      console.log('Successfully deleted');
+    } else {
+      console.log('Failed to delete');
+    }
+  };
+
+  return {
+    isDeleted,
+    setIsDeleted,
+    deleteBookmark
+  };
+};
